# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/analytics/popular-searches": {
            "get": {
                "tags": [
                    "analytics"
                ],
                "summary": "Get popular searches",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/trending": {
            "get": {
                "tags": [
                    "analytics"
                ],
                "summary": "Get trending skins",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period: 24h or 7d",
                        "name": "period",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Skin"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/skins": {
            "get": {
                "tags": [
                    "skins"
                ],
                "summary": "Get skins list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Weapon filter",
                        "name": "weapon",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Quality filter",
                        "name": "quality",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min price",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max price",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by: price, volume, name, updated",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order: asc, desc",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.SkinListResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/skins/chart/{slug}": {
            "get": {
                "tags": [
                    "skins"
                ],
                "summary": "Get price chart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skin slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Period: 24h, 7d, 30d, 90d, 1y, all",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PriceChartResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/skins/popular": {
            "get": {
                "tags": [
                    "skins"
                ],
                "summary": "Get popular skins",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Skin"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/skins/search": {
            "get": {
                "tags": [
                    "skins"
                ],
                "summary": "Search skins",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Skin"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/skins/{slug}": {
            "get": {
                "tags": [
                    "skins"
                ],
                "summary": "Get skin by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skin slug (e.g., awp_acheron_ft)",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Period: 24h, 7d, 30d, 90d, 1y, all",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.SkinDetailResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.PriceChartData": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
        "entity.PriceChartResponse": {
            "type": "object",
            "properties": {
                "avg_price": {
                    "type": "number"
                },
                "data_points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PriceChartData"
                    }
                },
                "max_price": {
                    "type": "number"
                },
                "min_price": {
                    "type": "number"
                },
                "period": {
                    "description": "24h, 7d, 30d, 90d, 1y, all",
                    "type": "string"
                },
                "skin_id": {
                    "type": "string"
                },
                "total_volume": {
                    "type": "integer"
                }
            }
        },
        "entity.PriceHistory": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "recorded_at": {
                    "type": "string"
                },
                "skin_id": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
...
        "entity.Skin": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "current_price": {
                    "type": "number"
                },
                "highest_price": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "last_updated": {
                    "type": "string"
                },
                "lowest_price": {
                    "type": "number"
                },
                "market_hash_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_change_24h": {
                    "type": "number"
                },
                "price_change_7d": {
                    "type": "number"
                },
                "quality": {
                    "type": "string"
                },
                "rarity": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "volume_24h": {
                    "type": "integer"
                },
                "weapon": {
                    "type": "string"
                }
            }
        },
        "entity.SkinDetailResponse": {
            "type": "object",
            "properties": {
                "price_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PriceHistory"
                    }
                },
                "skin": {
                    "$ref": "#/definitions/entity.Skin"
                },
                "statistics": {
                    "$ref": "#/definitions/entity.SkinStatistics"
                }
            }
        },
        "entity.SkinListResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "skins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Skin"
                    }
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "entity.SkinStatistics": {
            "type": "object",
            "properties": {
                "avg_price_30d": {
                    "type": "number"
                },
                "avg_price_7d": {
                    "type": "number"
                },
                "price_volatility": {
                    "type": "number"
                },
                "total_volume_7d": {
                    "type": "integer"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
...
    }
}